// src/enums/filters.enums.ts
export enum EStatus {
  ALL = "All",
  IN_STOCK = "In stock",
  LOW_STOCK = "Low stock",
  OUT_OF_STOCK = "Out of stock",
}

export enum EPromotion {
  ALL = "All",
  FEATURED = "Featured",
  STANDARD = "Standard",
}

export enum ESort {
  NEWEST = "Newest",
  PRICE_LOW = "Price low",
  PRICE_HIGH = "Price high",
  STOCK_LOW = "Stock low",
  STOCK_HIGH = "Stock high",
  NAME_ASC = "Name ascending",
  NAME_DESC = "Name descending",
}